Build fixes.

+++ include/common/StoreCommon.h
@@ -112,10 +112,10 @@ public:
 	bool Eof();
 	bool GetStatus();
 	void SetStatus(bool s);
-	int Read(void *Buffer, int Size, int Flags = 0);
-	int Write(const void *Buffer, int Size, int Flags = 0);
-	int ReadStr(char *Buf, int Size);
-	int WriteStr(char *Buf, int Size);
+	ssize_t Read(void *Buffer, ssize_t Size, int Flags = 0);
+	ssize_t Write(const void *Buffer, ssize_t Size, int Flags = 0);
+	ssize_t ReadStr(char *Buf, ssize_t Size);
+	ssize_t WriteStr(char *Buf, ssize_t Size);
 };
 
 class StorageItem
